requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
pandas==2.2.3
plotly==5.24.0
kaleido==0.2.1
openpyxl==3.1.5
pytz==2025.2
selenium==4.33.0
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.2
httpx<0.27.2
pytubefix==9.1.1
youtube-search-python==1.6.5
PyNaCl==1.5.0
psutil==7.0.0